+++ contracts/external/alpha/ISafeBoxETH.sol
@@ -4,7 +4,9 @@ pragma solidity >=0.8.0;
 import "@openzeppelin/contracts-upgradeable/token/ERC20/IERC20Upgradeable.sol";
 
 interface ISafeBoxETH is IERC20Upgradeable {
-  function cToken() external view returns (address);
-  function deposit() external payable;
-  function withdraw(uint amount) external;
+    function cToken() external view returns (address);
+
+    function deposit() external payable;
+
+    function withdraw(uint256 amount) external;
 }
