+++ contracts/external/compound/ICErc20.sol
@@ -10,5 +10,10 @@ import "./ICToken.sol";
  */
 interface ICErc20 is ICToken {
     function underlying() external view returns (address);
-    function liquidateBorrow(address borrower, uint repayAmount, ICToken cTokenCollateral) external returns (uint);
+
+    function liquidateBorrow(
+        address borrower,
+        uint256 repayAmount,
+        ICToken cTokenCollateral
+    ) external returns (uint256);
 }
