+++ contracts/external/compound/IRewardsDistributor.sol
@@ -12,13 +12,13 @@ interface IRewardsDistributor {
     function rewardToken() external view returns (address);
 
     /// @notice The portion of compRate that each market currently receives
-    function compSupplySpeeds(address) external view returns (uint);
+    function compSupplySpeeds(address) external view returns (uint256);
 
     /// @notice The portion of compRate that each market currently receives
-    function compBorrowSpeeds(address) external view returns (uint);
+    function compBorrowSpeeds(address) external view returns (uint256);
 
     /// @notice The COMP accrued but not yet transferred to each user
-    function compAccrued(address) external view returns (uint);
+    function compAccrued(address) external view returns (uint256);
 
     /**
      * @notice Keeps the flywheel moving pre-mint and pre-redeem
