+++ contracts/external/keep3r/Keep3rV1Oracle.sol
@@ -3,12 +3,14 @@ pragma solidity >=0.8.0;
 
 abstract contract Keep3rV1Oracle {
     struct Observation {
-        uint timestamp;
-        uint price0Cumulative;
-        uint price1Cumulative;
+        uint256 timestamp;
+        uint256 price0Cumulative;
+        uint256 price1Cumulative;
     }
 
     function factory() external pure virtual returns (address);
+
     mapping(address => Observation[]) public observations;
-    function observationLength(address pair) external view virtual returns (uint);
+
+    function observationLength(address pair) external view virtual returns (uint256);
 }
