+++ contracts/external/synthetix/ExchangeRates.sol
@@ -4,7 +4,7 @@ pragma solidity >=0.8.0;
 interface ExchangeRates {
     function effectiveValue(
         bytes32 sourceCurrencyKey,
-        uint sourceAmount,
+        uint256 sourceAmount,
         bytes32 destinationCurrencyKey
-    ) external view returns (uint value);
+    ) external view returns (uint256 value);
 }
