+++ contracts/external/synthetix/ISynthetix.sol
@@ -4,7 +4,7 @@ pragma solidity >=0.8.0;
 interface ISynthetix {
     function exchange(
         bytes32 sourceCurrencyKey,
-        uint sourceAmount,
+        uint256 sourceAmount,
         bytes32 destinationCurrencyKey
-    ) external returns (uint amountReceived);
+    ) external returns (uint256 amountReceived);
 }
